scripts: Add dsbize
This C file patches a homebrew into a DSBooter format, optionally with
an XOR-encrypted header. The binary header in that file is yet to be
documented, but it is suggested by asiekierka that it may be a variant
of Darkfader's old GBA boostrap.

This is needed by all M3-based flash cartridges.

With XOR encryption, every byte in the header is XORed with a key.
Known keys for specific cartridges are as follows:
- M3 DS Real jp : 0x07
- M3 DS Real eng: 0x12
- M3 DS Real gb : 0x23
- iTouchDS jp   : 0x37
- iTouchDS eng  : 0x32
- iTouchDS gb   : 0x33
- r4rts.com     : 0x72