notes: Ace3DS+ family can actually load A9/A7 anywhere in main RAM
The initial notes are not *entirely* correct; one doesn't need to use
the default ARM9 linkerscript in a given toolchain. Rather, for both
ARM9 and ARM7, anywhere in main RAM seems to be okay.

It appears the firmware may be loading the binaries in ARM9, which
would explain why loading the ARM7 in IWRAM does not work.