[DO NOT MERGE] Add hack for YSMenu
DSTT carts need a variable at 0x02FFFC24 to be zero or non-zero,
meaning the SD is non-SDHC/SDHC respectively. Without it, TTMenu and by
definition YSMenu will not be able to use its internal ROM loading
binary, ttpatch.

Unfortunately, this requires reinitializing the SD card to find out
whether the SD card has the HCS bit set; so import the source code for
the DSONE SDHC and use the functions here instead of the DLDI for
FatFs, and write the needful to that address.

DLDI is still required to load the YSMenu UI, so keep the patching
code.